+++ copy_trading_backend/src/main/java/com/xtq_ymt/copy_trading_backend/repository/MarketDataRepository.java
@@ -11,7 +11,7 @@ import org.springframework.data.jpa.repository.Modifying;
 
 import java.util.Date;
 import java.util.List;
-import java.math.BigDecimal;
+import java.math.BigDecimal; // 确保导入 BigDecimal
 
 @Repository
 public interface MarketDataRepository extends JpaRepository<MarketData, Long> {
@@ -32,17 +32,19 @@ public interface MarketDataRepository extends JpaRepository<MarketData, Long> {
     List<MarketData> findByVolumeLessThan(BigDecimal volume);
 
     // 查找市场波动率大于指定值的MarketData
-    List<MarketData> findByVolatilityGreaterThan(Double volatility);
+    List<MarketData> findByVolatilityGreaterThan(BigDecimal volatility); // 修改参数类型为 BigDecimal
 
     // 按时间戳降序排序，查找最近的MarketData（分页）
     Page<MarketData> findAllByOrderByTimestampDesc(Pageable pageable);
 
     // 自定义查询：查找在特定时间范围内的特定金融工具的历史记录
     @Query("SELECT m FROM MarketData m WHERE m.instrument = :instrument AND m.timestamp BETWEEN :startDate AND :endDate")
-    Page<MarketData> findHistoryByInstrumentAndDateRange(@Param("instrument") String instrument,
-                                                         @Param("startDate") Date startDate,
-                                                         @Param("endDate") Date endDate,
-                                                         Pageable pageable);
+    Page<MarketData> findHistoryByInstrumentAndDateRange(
+        @Param("instrument") String instrument,
+        @Param("startDate") Date startDate,
+        @Param("endDate") Date endDate,
+        Pageable pageable
+    );
 
     // 批量删除在指定时间范围内的MarketData
     @Modifying
